Ensure encrypted records don't underflow. Closes #23.
Records can underflow if the padding length is larger than the
plaintext length, and the record size is larger than the minimum
`rs` but smaller than the plaintext length. We should never emit
short records, since they'll fail to decrypt.

We *can* emit trailing padding-only blocks, but the draft discourages
this because it'll still leak size information. If we exhaust the
ciphertext and still have padding left, we fail encryption. This
matches the reference implementation's behavior.

This commit also adds some comments for the chunking logic, and uses
more consistent terminology: "plaintext-", "pad-", and "ciphertext-"
prefixed variables refer to their respective arrays, and "data-" is
the combined plaintext and padding.